+++ frontend/src/component/changeRequest/ProjectChangeRequests/ChangeRequestsTabs/ChangeRequestActionCell/ChangeRequestActionCell.tsx
@@ -1,0 +1,12 @@
+import { ArrowRight } from '@mui/icons-material';
+import { useTheme } from '@mui/system';
+import { TextCell } from '../../../../common/Table/cells/TextCell/TextCell';
+
+export const ChangeRequestActionCell = () => {
+    const theme = useTheme();
+    return (
+        <TextCell sx={{ textAlign: 'right' }}>
+            <ArrowRight sx={{ color: theme.palette.secondary.main }} />{' '}
+        </TextCell>
+    );
+};
