+++ src/lib/types/permissions.ts
@@ -37,4 +37,4 @@ export const MOVE_FEATURE_TOGGLE = 'MOVE_FEATURE_TOGGLE';
 export const CREATE_SEGMENT = 'CREATE_SEGMENT';
 export const UPDATE_SEGMENT = 'UPDATE_SEGMENT';
 export const DELETE_SEGMENT = 'DELETE_SEGMENT';
-export const SUGGEST_CHANGE = 'SUGGEST_CHANGE';
+export const CHANGE_REQUEST = 'CHANGE_REQUEST';
